extract logout logic for reusability
