+++ src/components/navigation/navbar/common/profile/ProfileMenuBody.vue
@@ -22,12 +22,7 @@
 
 <script lang="ts">
     import MenuItem from "../../common/MenuItem.vue";
-    import { useStore } from "../../../../../store/store";
-    import { MutationTypes } from "../../../../../store/mutation-types";
-    import Lecturer from "../../../../../api/api_models/user_management/Lecturer";
-    import Admin from "../../../../../api/api_models/user_management/Admin";
-    import Student from "../../../../../api/api_models/user_management/Student";
-    import { Role } from "../../../../../entities/Role";
+    import { logout } from "@/use/Logout";
 
     export default {
         name: "ProfileMenuBody",
@@ -36,11 +31,7 @@
         },
         setup() {
             function logOut() {
-                const store = useStore();
-                store.commit(MutationTypes.SET_LOGINDATA, { username: "", password: "" });
-                store.commit(MutationTypes.SET_USER, {} as Student | Lecturer | Admin);
-                store.commit(MutationTypes.SET_ROLE, Role.NONE);
-                store.commit(MutationTypes.SET_LOGGEDIN, false);
+                logout();
             }
 
             return {
